requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
plotly==5.17.0
pandas==2.1.1
matplotlib==3.7.2
gunicorn==21.2.0
